Move polygons to dedicated layer
